Add GoDaddy intermediate certificates
It appears that https://online.swagger.io is now using a certificate
issued by GoDaddy.  Unfortunately, this new certificate is also missing
the intermediate certificate necessary to validate it using the Mozilla
CA Bundle.  Add the necessary certs from
https://certs.godaddy.com/repository/:

GoDaddy Class 2 Certification Authority Root Certificate - G2
https://certs.godaddy.com/repository/gdroot-g2.crt
45:14:0B:32:47:EB:9C:C8:C5:B4:F0:D7:B5:30:91:F7:32:92:08:9E:6E:5A:63:E2:74:9D:D3:AC:A9:19:8E:DA

GoDaddy Secure Server Certificate (Intermediate Certificate) - G2
https://certs.godaddy.com/repository/gdig2.crt.pem
97:3A:41:27:6F:FD:01:E0:27:A2:AA:D4:9E:34:C3:78:46:D3:E9:76:FF:6A:62:0B:67:12:E3:38:32:04:1A:A6

Note that the Root Certificate is included in the Mozilla CA Bundle, so
it is not required in most cases.  It is included so the cert can be
verified even without a system trusted CA list.

Fixes: #38

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>