Clarify support for OpenAPI 2.0, not 3.0
@tim-digitised made a good observation in issue #45 that README.md is
misleading since the OpenAPI Specification which it links to currently
shows the OpenAPI Specification version 3.0, which is unsupported.  Fix
the link to point to the OpenAPI Specification repo and add text to
clarify supported versions and link to the open issue for 3.0 support in
swagger-api/validator-badge.

Also add the openapi2 tag to package.json so users can more easily
search for packages supporting 2.0 vs 3.0 in npm.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>